## Performance Enhancement

Optimization for the transaction pool's `get_all()` method which eliminates unnecessary function call overhead in hot paths during transaction retrieval operations. We replace indirect `get()` method calls with direct HashMap lookups, which reduces CPU overhead during batch operations like block building and transaction forwarding. Benchmarks show 4-17% faster operation with identical functionality and no regressive case.


## Benchmark Results

```
Testing small (10) transactions:
  Original:    1.14ms
  Optimized:   1.09ms     (4.3% improvement, 1.05x speedup)

Testing medium (100) transactions:
  Original:    9.59ms
  Optimized:   7.95ms     (17.1% improvement, 1.21x speedup)
  
Testing large (1000) transactions:
  Original:    75.32ms
  Optimized:   65.05ms    (13.6% improvement, 1.16x speedup)
```

## Changelist
`crates/transaction-pool/src/pool/txpool.rs`:
- Optimized get_all() method to use direct HashMap lookups instead of calling self.get(&tx) for each transaction hash

## Test Plan
- (Unit) All existing transaction pool tests pass without modification, confirming identical functionality
- (Integration) Verified the transaction pool works correctly with the new optimized implementation
- (Performance) Benchmarked both implementations showing consistent 4-17% performance improvements